requirements.txt
pyjwt[crypto]==2.8.0
cryptography==41.0.7
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4

